requirements.txt
Flask
google-generativeai
firebase-admin
pyttsx3
SpeechRecognition
gtts
requests
python-dotenv
